refactor: 입출금내역 view 분리, 옵저버 등록
