+++ src/observers/TransactionListObserver.js
@@ -1,0 +1,12 @@
+import { renderTransactionList } from "../components/transactionListView.js";
+
+export class TransactionListObserver {
+  constructor(itemsState) {
+    this.itemsState = itemsState;
+    itemsState.subscribe(this);
+  }
+
+  update() {
+    renderTransactionList();
+  }
+}
